Cache label layout in its own class
Cache event label in an EventLabel class, where cachedWidth and
cachedHeight are kept in sync with the StaticLayout.

Compared to the previous approach which stored cachedWidth and
cachedHeight in the EventChip, this avoids recalculating the label
layout when the EventChip bounds are cleared. As EventChip bounds are
updated much more often than the available width and height, this change
brings noticeable performance improvements on calendars with many
events, considering that recreating SpannableStrings is expensive.